ber.go: Protocol Validation and Defense
This commit attempts to address possible `panic` conditions due to access outside of the bounds of the slice.  There are specifically three changes:
  * A couple log messages are commented out due to them attempting to do an out of bounds view into the slice.
  * Existing protocol validation was fixed.  There was an attempt to guard against these conditions, however in many cases there was an off by one error, or other conditions had not been considered.
  * A PR from the parent fork which attempts to bring in a protocol fix for lengths between `0x80` and `0xFF` was also included as part of this: https://github.com/mozilla-services/pkcs7/pull/68

The logic for slice access has been reviewed carefully in an attempt to make sure we are defensive, but still allowing the nuance of the PKCS7 structure.  In addition extensive fuzzing has been conducted on these changes.